Changed the layout from FrameLayout to ConstraintLayout.
Added tools:text for all TextViews excluding labels.

Sorted the views within the XML so that they are in an order that more
closely matches the JSON.

Renamed the id for the "also known as" label to a more appropriate
value.

Renamed the id for the the "also known as" text view to a more
appropriate value.

Renamed the id for the "place of origin" text to a more appropriate
value.

Added an id for the "place of origin" label.

Added an id for the "description" label.

Renamed the id for the "description" text view to a more appropriate
value.

Added an id for the "ingredients" label.

Renamed the id for the "ingredients" text view to a more appropriate
value.

Added the appropriate constaints for all the views.